SSE: Create proper sessions
Previously, the SSE streamer in Tala had a notion of "session" that
wasn't really a session. It was only designed to handle one client at a
time. This could in some cases cause leaks between the streams of
different sessions.

This change creates a proper session, where session id, message counter,
the streamer thread etc. are kept, and it's the caller of the streaming
methods that is required to provide the session object on every call.

Change-Id: Ib9c627f56865bcdca8abba3dbdaa9e1dff21eb43