NLG: support move subsequences
Previously, the NLG matched the entire move sequence or the individual
moves.

This change adds the possibility to match the final moves in a sequence,
so that if the moves are ["a", "b", "c", "d"], and there is an NLG
entry for sequence ["b", "c", "d"], the utterance for that sequence will
be generated. The first move ("a") will not be generated.

This change meets the specific requirements of task 1231.

Change-Id: I9e251b9b60ec338dc4415e63dc17ac89ae53b79a