Add tala.utils.sse_client.StreamerQueue
The streamer queue is a queue for streams that runs the actual streaming
in a separate thread. It could also be seen as an increased level of
abstraction, as it handles most of the stream and session management.

It supports setting up a streamer client for a particular voice and
persona, streaming tokens, and taking down the client.

Change-Id: Ice33bff7666cc4c3d2b108a894d9448473b9822f