1491: Abort/succeed plan execution plan item
There is a need for a plan item that aborts the execution of a plan
other than the one that the plan item belongs to.

This change adds the possibility to use "signal_action_completion" and
"signal_action_failure" transitively, referencing the action of any
perform goal. Executing the completion plan item results in the action
being downdated without further ado. Executing the "failure" item is
that the action is aborted and that a failure report move with the
failure reason included is generated.

Also:
 - fix JSON:API version comparison
 - remove some garbage files
 - make sure that JSON:API version can be given as an argument when
   creating JSONAPIObject

Change-Id: I2da3048828c460bdae80050d1594b06062f6248f